refactor: extract tabs and tools to separate files
