+++ src/utils/tabs.ts
@@ -1,0 +1,15 @@
+const tabs = [
+    { name: "All", query: "" },
+    { name: "Formatters", query: "formatters" },
+    { name: "Converters", query: "converters" },
+    { name: "Generators", query: "generators" },
+    { name: "Validators", query: "validators" },
+    { name: "Encoders", query: "encoders" },
+    { name: "Decoders", query: "decoders" },
+    { name: "Minifiers", query: "minifiers" },
+    { name: "Beautifiers", query: "beautifiers" },
+    { name: "Parsers", query: "parsers" },
+    { name: "Others", query: "others" },
+  ]
+
+export default tabs
