+++ src/utils/tools.ts
@@ -1,0 +1,11 @@
+const tools = [
+    {
+      name: "Temporary Email Generator",
+      description: "Generate a disposable temporary email address.",
+      category: "generators",
+      link: "email",
+      id: "email",
+    },
+  ]
+
+export default tools
